// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface CryptoSignal {
    currentPrice: number;
    monthAvgVolume: number;
    name: string;
    priceChangePercentage24h: number;
    dayVolume: number;
    coinId: string;
    dayHigh: number;
    symbol: string;
    signalType: string;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface backendInterface {
    getAllSignals(): Promise<Array<CryptoSignal>>;
    scanMarkets(): Promise<Array<CryptoSignal>>;
    transform(input: TransformationInput): Promise<TransformationOutput>;
}